// src/app/interfaces/users.ts
export interface UsersInterface {
    _id: string;
    username: string;
    password: string;
    name: string;
    lastname: string;
    full_names: string;
    phone: string;
    mobile: string;
    email: string;
    creator: any;
    status: any;
    profile: any;
    career: any;
    created: Date;
    updated: Date;
}